pyproject.toml
# Main project metadata and configuration
[project]
name = "cvxrisk"
version = "0.0.0"
description = "Simple riskengine for cvxpy"
readme = "README.md"
requires-python = ">=3.11"
# List of core dependencies required for the package to function
dependencies = [
    "cvxpy-base>=1.7.2",    # Core convex optimization modeling framework
    "numpy>=2.3.0",        # Numerical computing library for array operations
    "pandas>=2.3.0",       # Data manipulation and analysis library
]
authors = [{name = "Thomas Schmelzer", email = "thomas.schmelzer@gmail.com"}]

# Project URLs for documentation, repository, etc.
[project.urls]
repository = "https://github.com/cvxgrp/cvxrisk"

# Optional dependencies that can be installed with extras (e.g., pip install cvxrisk[dev])
[project.optional-dependencies]
# Development dependencies for testing, linting, and documentation
dev = [
    "clarabel>=0.11.1",
    "pytest-cov>=6.3.0",   # Test coverage reporting
    "pytest>=9.0.1",       # Testing framework
    "pre-commit>=4.5.0",   # Git pre-commit hooks for code quality
    "cvxsimulator>=1.4.5", # Simulation framework for portfolio optimization
    "marimo>=0.17.0",     # Interactive notebook environment
    "polars>=1.34.0",      # Fast DataFrame library (alternative to pandas)
    #"micropip>=0.9.0",    # Package installer for WebAssembly/Pyodide
    "python-dotenv==1.2.1"
]

# Build system configuration
[build-system]
requires = ["hatchling"]  # Hatchling is used as the build backend
build-backend = "hatchling.build"

# Hatch configuration for building the package
[tool.hatch.build.targets.wheel]
packages = ["src/cvx"]  # Only include the cvx package in the wheel

# Deptry configuration for dependency checking
# Rule ignores for specific dependencies
[tool.deptry.per_rule_ignores]
DEP001 = ["cvx", "cvxpy", "sklearn"]  # Ignore missing imports (these are provided by other packages)
DEP002 = ["cvxpy-base", "scikit-learn", "mosek", "clarabel"]  # Ignore unused direct dependencies

[tool.deptry]
# see https://deptry.com/usage/#pep-621-dev-dependency-groups
pep621_dev_dependency_groups = ["dev"]
